+++ src/main.tsx
@@ -4,6 +4,7 @@ import App from "./App";
 import "./styles/theme.css";
 import "./styles/foundation.css";
 import "./styles/reference.css";
+import "./styles/preview-themes.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode><App /></React.StrictMode>,